requirements.txt
alembic==1.4.3
fastapi==0.61.1
FastAPI-SQLAlchemy==0.2.1
psycopg2==2.8.6
uvicorn==0.12.1
